presence [nfc]: Consolidate getPresence into the model file
